% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scatterplot.R
\name{make_scatterplot}
\alias{make_scatterplot}
\title{Create random scatterplot}
\usage{
make_scatterplot(tsv_file)
}
\arguments{
\item{tsv_file}{A string of the path of a tsv file with data to import}
}
\value{
A plot object from randmly chosen variables
}
\description{
Create random scatterplot
}
\examples{
\dontrun{
make_scatterplot("/path/to/dataset.tsv")
}
}
