% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dataset.R
\name{summarise_dataset}
\alias{summarise_dataset}
\title{Summarise a Dataset}
\usage{
summarise_dataset(tsv_file)
}
\arguments{
\item{tsv_file}{A string of the path of a tsv file with data to import}
}
\value{
A tibble with variables summarising average, min and max
}
\description{
Summarise a Dataset
}
\examples{
\dontrun{
summarise_dataset("/path/to/dataset.tsv")
}
}
